v0.2.6
Made methods virtual which were not converted into virtual after moving it to BaseSDK.
Configurable is an abstract class now.
GameStates and GameStates discarding data on the client side fixed by making GameState generic. Could have made the thing be parsed using a custom converter, but it looks like that would be using reflection, so avoided it for now.